pyproject.toml
[tool.poetry]
name = "solis"
version = "0.1.0"
description = "A tool to control Solis Hybrid Inverters"
authors = ["Dan Foster <dan@zem.org.uk>"]
license = "MIT"
readme = "README.md"
packages = [{include = "solis", from="src"}]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
asyncclick = "^8.1.3.4"
zem-pysolarmanv5 = "^2.5.0rc1"


[tool.poetry.group.dev.dependencies]
mypy = "^0.990"
flake8 = "^5.0.4"
black = "^22.10.0"
flake8-pyproject = "^1.1.0.post0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
"solis" = "solis.cli:main"